(#28): Make games that follow one another by 2 hours or less be consecutive.
Consecutive games disable the extra buffer given prior to the start of the game,
since it's assumed that the official is already on-site. In order for the game
to actually be a consecutive game, it must be a) within the time range of two
hours from the specified end date of the previous game, and b) at the same site
as the previous game.

Note: This commit only adds functionality that allows games that are entered
into Arbitrator together be checked for potential consecutive-ness. Any games
that are entered one-by-one (or a game that is entered into Arbitrator at a
later date from its preceding counterpart) will not be checked for the
possibility of consecutive game linkage.