Make back navigation re-initialize the google calendar list.
This caches the GoogleArbitratorClient within the UIManager class. In doing
so, we provide a method for obtaining it (possibly with a callback so that
initialization can occur, followed by the necessary commands to execute on
the new client object).

When the special 'main' view is detected during loadContent(), the google
client is retrieved, and the userId and calendar are populated
appropriately.

Fixes #49.