Rework the PreferenceStore to make it an actual singleton object.
Previously, the preference store would create a new instance of itself
every time new PreferenceStore() was run (a duh). This is suboptimal in
tests, though, because (as part of this change) we needed to remove the
write to disk of preference objects, so as not to interfere with
development/production preferences.

Refs #64.