Make LocationService module utilize the official google maps node module.
This re-enables the address prediction on the location settings view by
refactoring LocationService to utilize the official google maps node
module. Unfortunately, the previous module that was being used in the web
application doesn't work within an electron app. Further complicating this
is the fact that the @google/maps module doesn't give any method of
actually populating the dropdown list (this is only available in the web
app javascript library). Even worse, the <datalist> element doesn't work
within electron right now.

To solve all of these problems, the typeahead module was used to display
the list, and the individual predictions of Google Places comes from the
@google/maps module.

Fixes #75.