Make the ArbiterSports connection check autonomous.
Previous to this commit, the connection window (in settings) for
ArbiterSports was not automatic. Now, it checks for a valid connection and
gives a response if the connection is invalid or valid, after closing the
window.

This also creates the window from the main process, rather than a separate
renderer process.

Refs #70.