Separate gulp tasks for release into a package, gulp-ssh-deploy.
Arbitrator now depends on a separate npm module, gulp-ssh-deploy. The
purpose of this package is to add the necessary release tasks to the gulp
configuration. Only the configuration of this needs to be passed to the
gulp-ssh-deploy module, which happens in tasks/release.js now, rather than
the manual setup of these tasks.

Fixes #87.