Make UIManager check if sublocation and address are default before setting.
Previous to this commit, UIManager would overwrite both address and
sublocation for each location upon which the 'Set' button was clicked,
regardless of whether the user entered data into that field. This commit
makes the UIManager slightly smarter, in that, if a user has not entered
anything into a particular field (e.g. address), but something was
previously entered/saved (i.e. the placeholder for that field is
non-default), then it will save that data instead of whatever is returned
by val() for that field.

Fixes #78.