Add the ability to have titles with spaces for nav drawer entries.
If a title should have a space in it, the space should be included in the
data-item attribute of the nav drawer entry element. This will be converted
to a dash (-) for partial html file specification.

This also removes the svg icons for the nav drawer items and replaces it
with font-awesome variants.

Refs #21.