Rename profile classes to more suitable and intuitive names.
Renames GameAgeProfile to LeagueProfile and GameAgeLevel to
GameClassificationLevel. Now, the classification can be anything (e.g.
'Boys'), and isn't restricted to an age classsification. Similarly, the
GameAgeProfile doesn't just have settings for game ages. It does _right
now_, but this will change in the future, as all league-specific settings
(such as alias) will be placed in this overarching container.

Refs #21.